Create handlers and config pkgs
