+++ jobs/aws-batch-jobs.go
@@ -15,7 +15,7 @@ import (
 )
 
 type AWSBatchJob struct {
-	ctx         context.Context
+	Ctx         context.Context
 	ctxCancel   context.CancelFunc
 	UUID        string `json:"jobID"`
 	AWSBatchID  string
@@ -101,15 +101,15 @@ func (j *AWSBatchJob) ProviderID() string {
 func (j *AWSBatchJob) Equals(job Job) bool {
 	switch jj := job.(type) {
 	case *AWSBatchJob:
-		return j.ctx == jj.ctx
+		return j.Ctx == jj.Ctx
 	default:
 		return false
 	}
 }
 
 func (j *AWSBatchJob) Create() error {
-	ctx, cancelFunc := context.WithCancel(j.ctx)
-	j.ctx = ctx
+	ctx, cancelFunc := context.WithCancel(j.Ctx)
+	j.Ctx = ctx
 	j.ctxCancel = cancelFunc
 
 	batchContext, err := controllers.NewAWSBatchController(os.Getenv("AWS_ACCESS_KEY_ID"), os.Getenv("AWS_SECRET_ACCESS_KEY"), os.Getenv("AWS_DEFAULT_REGION"))
@@ -121,7 +121,7 @@ func (j *AWSBatchJob) Create() error {
 	log.Debug("j.JobDef | ", j.JobDef)
 	log.Debug("j.JobQueue | ", j.JobQueue)
 	log.Debug("j.JobName  | ", j.JobName)
-	aWSBatchID, err := batchContext.JobCreate(j.ctx, j.JobDef, j.JobName, j.JobQueue, j.Cmd, j.EnvVars)
+	aWSBatchID, err := batchContext.JobCreate(j.Ctx, j.JobDef, j.JobName, j.JobQueue, j.Cmd, j.EnvVars)
 	if err != nil {
 		j.HandleError(err.Error())
 		return err
@@ -226,7 +226,7 @@ func (j *AWSBatchJob) GetSizeinCache() int {
 	linkData := int(unsafe.Sizeof(j.Links))
 
 	totalMemory := cmdData + messageData + linkData +
-		int(unsafe.Sizeof(j.ctx)) +
+		int(unsafe.Sizeof(j.Ctx)) +
 		int(unsafe.Sizeof(j.ctxCancel)) +
 		int(unsafe.Sizeof(j.UUID)) + len(j.UUID) +
 		int(unsafe.Sizeof(j.AWSBatchID)) + len(j.AWSBatchID) +
